Fix @sys/lookup score failing output-schema validation
score was emitted as float32; the schema validator only accepts float64
for "type":"number", matching what encoding/json produces. Cast at the
handler boundary where float32 becomes map[string]any.

+++ kernel/call.go
@@ -358,7 +358,7 @@ func (k *Kernel) executeLookup(ctx context.Context, args map[string]any) (map[st
 			"name":         r.Action.Name,
 			"owner_handle": r.OwnerHandle,
 			"description":  r.Action.Description,
-			"score":        r.Score,
+			"score":        float64(r.Score),
 		}
 	}
 	return map[string]any{"results": items}, nil
